chore(hook): add cfn_nag support
Add support for Stelligent's cfn_nag tool (stelligent/cfn_nag), to
identify patterns that may indicate insecure infrastructure.

Because the cfn_nag repository does not have a .pre-commit-hooks.yaml
file, pre-commit cannot simply reference the repository and create the
Git hook. To circumvent this, we use a local hook which pulls a Docker
image provided by the project, and run the tool from within the
container. This is also convenient because cfn_nag runs on Ruby, which
would require the developer to install an additional dependency to run
the tool.